Fix hero image on 'About' page
The hero isn't sized to the full width of the browser for larger browser sizes. Looking further, the implementation of this in the blog's about page has the correct implementation of allowing the image to resize while cutting off what's visible at a maximum of 500px. Fixing the site's About page to match this.

* Assign hero image's container to a maximum height with the overflow hidden.

* Remove the maximum height on the hero image.

* Assign the width of the hero image to be full-width of the browser.