Update Mixpanel user information on sign in.
The issue is probably not updating user information, because we actually do it within
logged_in event. I think problem was that we did not alias anonymous
user to the one that logged in. This can cause issue when user enters
the website from different browser, clears cookies etc. Without aliasing
user, if we search in new browser for X, and then log in, the person who
conducted search for X would stay anonymous. With alias, we will know
that it wasn't anonymous person, but the one that has just logged in.

https://trello.com/c/aflkwY3n/247-update-mixpanel-user-information-on-sign-in