Locations page crashes IE8.
The issue that IE crashes occurs randomly, but pretty frequently. I
think that I have hunted down the line of code that causes issue - after
deleting it, I was not able to reproduce the error after few tries and
after re-adding it, the issue was observable again. Let's test on
staging if this helps.

https://trello.com/c/DpTWr3dN/315-locations-page-crashes-ie8