NM-382 Add CSV file endpoint for platform contacts and demo-requests.
The obscured routes replace the 'data-clips' functionality we lost when
we switched from heroku. Currently, the sales team uses the dataImport
function on google spreadsheets to import their sales contacts; it
requires a public url and connot do auth of any kind. Heroku's approach,
which I copied, is to use obscured urls behind ssl. Since the ssl
handshake happens before the url is sent, it is never sent in the clear
and is not susceptible to replay attacks.