NM-431 Address not being applied correctly to listing
* fix bug on space wizard page that fields related to address had wrong
 ids (location_<attribute> was expected, but was
 user_companies_attributes_0_locations_attributes_0_<attribute>). Due to
 this, we always used backend geocoder to geolocate address. For input
351 Paseo Nuevo, Santa Barbara, CA it returns formatted address Mueseum
of Contemporary Arts Santa Barbara, 653 paseo Nuevo, Santa Barbara, that
explains the issue.
* pick first suggestion from places Autocomplete if user fails to
choose one - the reason for that is that if user fails to pick
suggestion, we will fallback to our backend geocoder, which would lead
to the exact same bug. this requires making query to geocoder ( I
guess... couldn't find other solution ), so I used setTimeout function
to detect if user picked something or not ( google maps's event is fired
after blur, so we need to wait some time to know what happened )
* removes setting cityAddress for location's address - that should fix
 the issue with not being possible to enter address user wants